Add Turbo script for creating development prompts
