+++ packages/core/src/prompts/lesson-assistant/parts/generateResponse.ts
@@ -1,0 +1,3 @@
+export const generateResponse =
+  () => `RULES FOR HOW YOU SHOULD FORMAT YOUR RESPONSE
+You should respond with a valid JSON document where each key of the object corresponds with the keys of the lesson plan. The value of each key should be the content for that part of the lesson plan. The content should obey the schema I have set you for generating lesson plans.`;
