+++ packages/core/src/prompts/lesson-assistant/parts/index.ts
@@ -1,0 +1,12 @@
+export { americanToBritish } from "./americanToBritish";
+export { basedOn } from "./basedOn";
+export { body } from "./body";
+export { context } from "./context";
+export { endingTheInteraction } from "./endingTheInteraction";
+export { generateResponse } from "./generateResponse";
+export { interactive } from "./interactive";
+export { rag } from "./rag";
+export { schema } from "./schema";
+export { signOff } from "./signOff";
+export { task } from "./task";
+export { yourInstructions } from "./yourInstructions";
